+++ CHANGELOG.md
@@ -5,6 +5,7 @@
 - Added the “‘New’ Button Label” field setting. ([#159](https://github.com/craftcms/ckeditor/discussions/159))
 - Copy/pasting nested entry cards now duplicates the nested entries. ([#186](https://github.com/craftcms/ckeditor/issues/186), [#192](https://github.com/craftcms/ckeditor/pull/192))
 - Fixed a bug where it was possible to copy/paste nested entry cards between CKEditor fields. ([#192](https://github.com/craftcms/ckeditor/pull/192))
+- Fixed a bug where additional revisions were getting created when saving entries with CKEditor-nested entries.
 
 ## 4.0.0-beta.10 - 2024-03-12